[fix] : chatbot page
