+++ src/App.css
@@ -865,22 +865,34 @@ input[type="password"] {
 .chatbot-font {
   font-size: 17px;
   color: #606060;
-  padding-top: 17px;
-  padding-left: 50px;
-  padding-right: 500px;
+  margin-top: 17px;
+  padding: 10px;
   display: flex;
   justify-content: flex-start;
   text-align: start;
+  background-color: #d7e9fa;
+  width: fit-content;
+  border-radius: 10px;
+  max-width: 80%;
 }
 
 .chatbot-user-font {
+  /* position: absolute;
+  left: 100%; */
+  width: 95%;
   font-size: 17px;
   color: #606060;
-  padding-top: 17px;
-  padding-left: 500px;
-  padding-right: 50px;
+  margin-top: 17px;
+  padding: 10px;
   display: flex;
   justify-content: flex-end;
+}
+.chatbot-user-font > div {
+  width: fit-content;
+  padding: 10px;
+  background-color: #fffaaf;
+  border-radius: 10px;
+  max-width: 80%;
   text-align: start;
 }
