+++ src/App.jsx
@@ -24,6 +24,9 @@ import PostsPerCountry from "./component/PostsPerCountry";
 import DetailPost from "./component/DetailPost";
 import MyTripMoreInformation from "./component/MyTripMoreInformation";
 import MyTripLoveMoreInformation from "./component/MyTripLoveMoreInformation";
+import Landmark from "./component/Landmark";
+import Place from "./component/Place";
+import Plan from "./component/Plan";
 
 function App() {
   return (
@@ -37,7 +40,11 @@ function App() {
           <Route path="/signin" element={<SignIn />}></Route>
           <Route path="/mypage" element={<Mypage />}></Route>
           <Route path="/sharetrip" element={<ShareTrip />}></Route>
-          <Route path="/chatbot" element={<Chatbot />}></Route>
+          <Route path="/chatbot" element={<Chatbot />}>
+            <Route path="landmark-recommend" element={<Landmark />}></Route>
+            <Route path="place-recommend" element={<Place />}></Route>
+            <Route path="plan-recommend" element={<Plan />}></Route>
+          </Route>
           <Route path="/mytrip" element={<Mytrip />}></Route>
           <Route path="/findLoginId" element={<FindLoginId />}></Route>
           <Route path="/findPassword" element={<FindPassword />}></Route>
